requirements.txt
viam-sdk~=0.40.0